+++ src/MicroElements.NSwag.FluentValidation/ReflectionExtensions.cs
@@ -10,14 +10,11 @@ namespace MicroElements.NSwag.FluentValidation
     /// <summary>
     /// Reflection extensions
     /// </summary>
-    public static class ReflectionExtension
+    public static class ReflectionExtensions
     {
         /// <summary>
         /// Is sub class of generic type
         /// </summary>
-        /// <param name="child"></param>
-        /// <param name="parent"></param>
-        /// <returns></returns>
         public static bool IsSubClassOfGeneric(this Type child, Type parent)
         {
             if (child == parent)
@@ -39,21 +36,30 @@ namespace MicroElements.NSwag.FluentValidation
                 if (parent == cur || (isParameterLessGeneric && cur.GetInterfaces()
                                                                    .Select(i => GetFullTypeDefinition(i))
                                                                    .Contains(GetFullTypeDefinition(parent))))
+                {
                     return true;
+                }
                 else if (!isParameterLessGeneric)
+                {
                     if (GetFullTypeDefinition(parent) == cur && !cur.IsInterface)
                     {
                         if (VerifyGenericArguments(GetFullTypeDefinition(parent), cur))
+                        {
                             if (VerifyGenericArguments(parent, child))
                                 return true;
+                        }
                     }
                     else
-                        foreach (var item in child.GetInterfaces()
-                                                  .Where(
-                                                      i => GetFullTypeDefinition(parent) == GetFullTypeDefinition(i)
-                                                  ))
+                    {
+                        foreach (var item in child
+                                     .GetInterfaces()
+                                     .Where(i => GetFullTypeDefinition(parent) == GetFullTypeDefinition(i)))
+                        {
                             if (VerifyGenericArguments(parent, item))
                                 return true;
+                        }
+                    }
+                }
 
                 child = child.BaseType;
             }
@@ -72,12 +78,18 @@ namespace MicroElements.NSwag.FluentValidation
             Type[] parentArguments = parent.GetGenericArguments();
 
             if (childArguments.Length == parentArguments.Length)
+            {
                 for (var i = 0; i < childArguments.Length; i++)
+                {
                     if (childArguments[i].Assembly != parentArguments[i].Assembly ||
                         childArguments[i].Name != parentArguments[i].Name ||
                         childArguments[i].Namespace != parentArguments[i].Namespace)
+                    {
                         if (!childArguments[i].IsSubclassOf(parentArguments[i]))
                             return false;
+                    }
+                }
+            }
 
             return true;
         }